+++ mrv2/lib/mrvGL/mrvTimelineViewportEvents.cpp
@@ -67,6 +67,14 @@ namespace mrv
             Fl::remove_timeout((Fl_Timeout_Handler)laserFade_cb, data);
             // Remove shape from list
             data->annotation->remove(s);
+            if (data->annotation->empty())
+            {
+                auto player = getTimelinePlayer();
+                if (!player)
+                    return;
+                player->removeAnnotation(data->annotation);
+                updateUndoRedoButtons();
+            }
             // Remove callback data
             delete data;
         }
@@ -532,7 +540,7 @@ namespace mrv
                     shape->laser = laser;
                     shape->pts.push_back(pnt);
                     annotation->push_back(shape);
-                    _createAnnotationShape();
+                    _createAnnotationShape(laser);
                     break;
                 }
                 case ActionMode::kErase:
@@ -543,7 +551,7 @@ namespace mrv
                     shape->soft = softBrush;
                     shape->pts.push_back(pnt);
                     annotation->push_back(shape);
-                    _createAnnotationShape();
+                    _createAnnotationShape(false);
                     break;
                 }
                 case ActionMode::kArrow:
@@ -559,7 +567,7 @@ namespace mrv
                     shape->pts.push_back(pnt);
                     shape->pts.push_back(pnt);
                     annotation->push_back(shape);
-                    _createAnnotationShape();
+                    _createAnnotationShape(laser);
                     break;
                 }
                 case ActionMode::kCircle:
@@ -573,7 +581,7 @@ namespace mrv
                     shape->radius = 0;
 
                     annotation->push_back(shape);
-                    _createAnnotationShape();
+                    _createAnnotationShape(laser);
                     break;
                 }
                 case ActionMode::kRectangle:
@@ -589,7 +597,7 @@ namespace mrv
                     shape->pts.push_back(pnt);
                     shape->pts.push_back(pnt);
                     annotation->push_back(shape);
-                    _createAnnotationShape();
+                    _createAnnotationShape(laser);
                     break;
                 }
                 case ActionMode::kText: